% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{cell.centers}
\alias{cell.centers}
\title{Get center of grid cells for a retangular grid}
\usage{
cell.centers(xgrid, ygrid = NULL)
}
\arguments{
\item{xgrid}{Interfaces in x-direction}

\item{ygrid}{Interfaces in y-direction (if 2D)}
}
\description{
cell.centers takes a rectangular grid and returns coordinates of the cell centers as a field
}
