% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdetools.R
\name{CrossVariation}
\alias{CrossVariation}
\title{Cross-variation of two stochastic processes}
\usage{
CrossVariation(X, Y)
}
\arguments{
\item{X}{numeric vector containing the one process}

\item{Y}{numeric vector, same length as X, containing the other process}
}
\value{
A numeric vector, same length as X, giving the quadratic cross-varation as function of time
}
\description{
Cross-variation of two stochastic processes
}
\examples{
## Quadratic variation of Brownian motion
times <- seq(0,10,0.01)
B <- rBM(times)
plot(times,CrossVariation(B,B),type="l")
abline(0,1,lty="dashed")

## Verifying Ito's formula
X <- -times + B
h <- dhdx <- dh2dx2 <- function(x) exp(x)
Y <- h(X)
plot(times,Y,type="l")

Yi <- h(X[1]) + itointegral(dhdx(X),X) + 0.5*itointegral(dh2dx2(X),CrossVariation(X,X))
lines(times,Yi,col="blue")

}
