% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdetools.R
\name{itointegral}
\alias{itointegral}
\title{Ito integral of a stochastic process w.r.t. another}
\usage{
itointegral(G, B)
}
\arguments{
\item{G}{numeric vector containing the integrator}

\item{B}{numeric vector, same length as G, containing the integrand}
}
\value{
A numeric vector, same length as G, giving the "running integral", i.e. the Ito integral as a function of the upper limit.
}
\description{
Ito integral of a stochastic process w.r.t. another
}
\examples{
## Integration of Brownian motion w.r.t. itself
times <- seq(0,10,0.01)
BM <- rBM(times)
I <- itointegral(BM,BM)
matplot(times,cbind(I,0.5*BM^2-0.5*times),type="l",xlab="Time",ylab="Ito integral",
         main="Integral of B.M. w.r.t itself")

}
