% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdetools.R
\name{lyap}
\alias{lyap}
\title{Algebraic Lyapunov equation A*X+X*t(A)+Q=0}
\usage{
lyap(A, Q)
}
\arguments{
\item{A}{A quadratic matrix without eigenvalues on the imaginary axis}

\item{Q}{A symmetric matix of same dimension as A}
}
\value{
X A symmetric matrix of same dimension as A
}
\description{
Algebraic Lyapunov equation A*X+X*t(A)+Q=0
}
\details{
If A is asymptotically stable, Q is positive semidefinite and the pair (A,Q) is controllable, then X will be positive definite. Several similar results exist.
The implementation uses vectorization and kronecker products and does not employ sparsity, so is only suitable for small systems.
}
\examples{
# A scalar example
(lyap(-1,1))
# A harmonic oscillator
(lyap(array(c(0,-1,1,-0.1),c(2,2)),diag(c(0,1))))

}
