% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{pack.field}
\alias{pack.field}
\title{Convert a tabulated function on the plane to a vector}
\usage{
pack.field(phi)
}
\arguments{
\item{phi}{Array of probabilities for a 2D grid}
}
\description{
pack.field takes an array (e.g., of probabilities) and produces a vector.
}
