% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{QuasiStationaryDistribution}
\alias{QuasiStationaryDistribution}
\title{Compute the quasi-stationary distribution for a terminating Continuous Time Markov Chain}
\usage{
QuasiStationaryDistribution(G, max.iter = 20, tol = 1e-16)
}
\arguments{
\item{G}{Sub-generator of a CTMC: A quadratic matrix with non-negative off-diagonal elements and non-positive row sums}
}
\value{
A list containing:
vector, a vector with non-negative elements, containing the stationary distribution
value, the corresponding eigenvalue.
}
\description{
QuasiStationaryDistribution compute the quasi-stationary distribution phi for a Continuous Time
Markov Chain given by a sub-generator G, which is given by the equations

  phi %*% G == -mu*phi, sum(phi) == 1
}
\examples{
xgrid <- seq(0,1,0.01)
G <- fvade(u=function(x)5,D=function(x)1,xgrid=xgrid,bc="a")
sol <- QuasiStationaryDistribution(G)
plot(cell.centers(xgrid),sol$vector/diff(xgrid),main=paste("Decay rate",-sol$value))
}
