% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdetools.R
\name{rBM}
\alias{rBM}
\title{Simulate a sample path of Brownian motion}
\usage{
rBM(times, sigma = 1, B0 = 0, u = 0)
}
\arguments{
\item{times}{numeric vector of time points where the Brownian motion should be simulated}

\item{sigma}{numeric scalar, noise level in the Brownian motion, defaults to 1}

\item{B0}{numeric scalar, initial condition (at time=0) for the Brownian motion}

\item{u}{numeric scalar, drift}
}
\value{
numeric vector, of same length as times, containing the simulated value of the Brownian motion
}
\description{
Simulate a sample path of Brownian motion
}
\examples{
times <- 0:10
plot(times,rBM(times),type="b",xlab="Time",ylab="Brownian motion")

}
\seealso{
\code{\link{rvBM}} for a vectorized version
}
