% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdetools.R
\name{rBrownianBridge}
\alias{rBrownianBridge}
\title{Simulate a Brownian bridge}
\usage{
rBrownianBridge(times, B0T = c(0, 0), sigma = 1)
}
\arguments{
\item{times}{numeric vector of time points where the Brownian motion should be simulated}

\item{B0T}{initial and terminal condition; the fixed value of the Brownian motion at the first and last time points}

\item{sigma}{scalar; the noise level in the Brownian motion.}
}
\value{
a numeric vector, length(times), containing a sample path
}
\description{
Simulate a Brownian bridge
}
\examples{
times <- 10:20
BB <- rBrownianBridge(times,c(10,-10),2)
plot(times,BB,type="b")

}
