% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{StationaryDistribution}
\alias{StationaryDistribution}
\title{Compute the stationary distribution for a Continuous Time Markov Chain}
\usage{
StationaryDistribution(G)
}
\arguments{
\item{G}{Generator of a CTMC: A quadratic matrix with non-negative off-diagonal elements and zero row sums}
}
\value{
phi, a vector with as many elements as rows in G, containing the stationary distribution.
}
\description{
StationaryDistribution compute the stationary distribution phi for a Continuous Time
Markov Chain given by a generator G, which is given by the equations

  phi %*% G == 0, sum(phi) == 1

It is assumed that these two equations specify phi uniquely, i.e. that G is the generator
of an ergodic chain.
}
\examples{
G <- array(c(-2,1,2,-1),c(2,2))
phi <- StationaryDistribution(G)

}
