% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{unpack.field}
\alias{unpack.field}
\title{Convert a vector to a function on the plane}
\usage{
unpack.field(phi, nx, ny)
}
\arguments{
\item{phi}{Tabulated function}

\item{nx}{Number of grid cells in the x direction}

\item{ny}{Number of grid cells in the y direction}
}
\description{
unpack.field takes a vector and produces a 2D array.
}
\details{
length(phi) must equal nx * ny.
}
